Mock Jadex Platform

+++ wot-servient-binding-jadex/src/main/java/city/sane/wot/binding/jadex/JadexProtocolServer.java
@@ -1,15 +1,10 @@
 package city.sane.wot.binding.jadex;
 
 import city.sane.wot.binding.ProtocolServer;
+import city.sane.wot.binding.ProtocolServerException;
 import city.sane.wot.thing.ExposedThing;
 import com.typesafe.config.Config;
-import com.typesafe.config.ConfigObject;
-import jadex.base.IPlatformConfiguration;
-import jadex.base.PlatformConfigurationHandler;
-import jadex.base.Starter;
 import jadex.bridge.IExternalAccess;
-import jadex.bridge.service.search.ServiceQuery;
-import jadex.bridge.service.types.cms.CreationInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -27,39 +22,26 @@ public class JadexProtocolServer implements ProtocolServer {
     private static final Logger log = LoggerFactory.getLogger(JadexProtocolServer.class);
 
     private final Map<String, ExposedThing> things = new HashMap<>();
-    private final IPlatformConfiguration platformConfig;
+    private final JadexProtocolServerConfig config;
     private IExternalAccess platform;
     private ThingsService thingsService;
 
-    public JadexProtocolServer(Config config) {
-        platformConfig = PlatformConfigurationHandler.getDefault();
-        ConfigObject objects = config.getObject("wot.servient.jadex.server");
-        objects.forEach((key, value) -> platformConfig.setValue(key, value.unwrapped()));
+    public JadexProtocolServer(Config wotConfig) {
+        config = new JadexProtocolServerConfig(wotConfig);
+    }
+
+    JadexProtocolServer(JadexProtocolServerConfig config) {
+        this.config = config;
     }
 
     @Override
     public CompletableFuture<Void> start() {
         log.info("JadexServer is starting Jadex Platform");
 
-        CompletableFuture<IExternalAccess> createPlatform = FutureConverters.fromJadex(Starter.createPlatform(platformConfig));
-        CompletableFuture<IExternalAccess> createThingsAgent = createPlatform.thenCompose(agent -> {
-            CreationInfo info = new CreationInfo()
-                    .setFilenameClass(ThingsAgent.class)
-                    .addArgument("things", things);
-            return FutureConverters.fromJadex(agent.createComponent(info));
+        return config.createPlatform(things).thenAccept(result -> {
+            platform = result.first();
+            thingsService = result.second();
         });
-
-        CompletableFuture<ThingsService> searchThingsService = createThingsAgent.thenCompose(agent -> FutureConverters
-                .fromJadex(agent.searchService(new ServiceQuery<>(ThingsService.class))));
-
-        CompletableFuture<ThingsService> start = createPlatform
-                .thenCombine(searchThingsService, (ia, service) -> {
-                    platform = ia;
-                    thingsService = service;
-                    return thingsService;
-                });
-
-        return start.thenApply(r -> null);
     }
 
     @Override
@@ -80,7 +62,7 @@ public class JadexProtocolServer implements ProtocolServer {
         things.put(thing.getId(), thing);
 
         if (platform == null) {
-            return CompletableFuture.failedFuture(new Exception("Unable to expose thing before JadexServer has been started"));
+            return CompletableFuture.failedFuture(new ProtocolServerException("Unable to expose thing before JadexServer has been started"));
         }
 
         CompletableFuture<IExternalAccess> expose = FutureConverters.fromJadex(thingsService.expose(thing.getId()));
@@ -92,7 +74,12 @@ public class JadexProtocolServer implements ProtocolServer {
     public CompletableFuture<Void> destroy(ExposedThing thing) {
         things.remove(thing.getId());
 
-        return FutureConverters.fromJadex(thingsService.destroy(thing.getId()));
+        if (thingsService != null) {
+            return FutureConverters.fromJadex(thingsService.destroy(thing.getId()));
+        }
+        else {
+            return CompletableFuture.completedFuture(null);
+        }
     }
 
 }
